Split Memory class into Storage and TaskList class, add Ui class
Memory class was in charge of both manupilating the stored tasks as well as saving the tasks to external file.

Now storage class handles the saving and reading of data from the external file.
TaskLists handles the manipulation, including adding, removing and marking etc.
Ui handles all interaction with user.
