+++ src/main/java/Event.java
@@ -15,12 +15,20 @@ public class Event extends Task {
         this.duration = duration;
     }
 
+    /**
+     * Converts the event to string representation to be stored in text file
+     * @return a string
+     */
     @Override
     public String getDescription() {
         String status = super.getStatus() ? "T" : "F";
         return "E | " + status + " | " + super.toString() + " | " + duration + "\n";
     }
 
+    /**
+     * Converts the event to string representation for user
+     * @return a string
+     */
     @Override
     public String toString() {
         return "[E]" + super.getStatusIcon() + " " + super.toString() + "(at: " + this.duration + ")";
